models(feature): implement models
    - add relations
    - add protects get events

[Delvers #ESL-2]
