+++ prisma/schema.prisma
@@ -13,6 +13,28 @@ model User {
   username      String?
   email         String?   @unique
   password      String?
+  eventTypes    EventTypes[]
   createdAt     DateTime  @default(now()) @map(name: "created_at")
   updatedAt     DateTime  @updatedAt @map(name: "updated_at")
+}
+
+model EventTypes {
+  id            Int       @default(autoincrement()) @id
+  user          User @relation(fields: [id], references: [id])
+  name          String
+  duration      Int
+  events        Events[]
+}
+
+model Events {
+  id            Int @default(autoincrement()) @id
+  eventType     EventTypes @relation(fields: [id], references: [id])
+  attendies     Attendees[]
+}
+
+model Attendees{
+  id           Int @default(autoincrement()) @id
+  name         String
+  email        String? @unique
+  event        Events @relation(fields: [id], references: [id])
 }
